+++ geir_magi_mat/src/controllers/NavMenuController.js
@@ -1,5 +1,5 @@
 import { model } from "../models/index.js";
-import { closeMenu, openMenu } from "../views/NavMenuView.js";
+import { closeMenu, openMenu } from "../views/ui/NavMenuView.js";
 
 export const handleToggleMenu = () => {
     model.app.navMenuOpen = !model.app.navMenuOpen;