+++ geir_magi_mat/src/views/ui/NavBarView.js
@@ -1,0 +1,11 @@
+export function NavBarView() {
+    return /* html */ `
+        <nav>
+            <div class="empty-placeholder" id="dont-remove-plz"></div>
+            <a href="/" data-link id="logo">
+                <h1>Middags-Magi</h1>
+            </a>
+            <div onclick="handleToggleMenu()"><i class="fa-solid fa-bars fa-xl"></i></div>
+        </nav>
+    `;
+}