+++ geir_magi_mat/src/views/view.js
@@ -1,28 +1,17 @@
 import { model } from "../models/index.js";
-import { FrontPageView } from "./FrontPageView.js";
-import { ExploreRecipesView } from "./ExploreRecipesView.js";
-import { NewRecipeView } from "./NewRecipeView.js";
-import { setupNewRecipePage } from "../views/NewRecipeView.js";
-import { myRecipesView } from "./MyRecipesView.js";
-import { FullRecipeView } from "./FullRecipeView.js";
-import { editRecipeView } from "./EditRecipeView.js";
-import { navMenu } from "./NavMenuView.js";
-
-const navBar = /* html */ `
-    <nav>
-        <div class="empty-placeholder" id="dont-remove-plz"></div>
-        <div onclick="switchPage('FrontPage')" id="logo">
-            <h1>Middags-Magi</h1>
-        </div>
-        <div onclick="handleToggleMenu()"><i class="fa-solid fa-bars fa-xl"></i></div>
-    </nav>
-`;
-
-const footer = /* html */ ` <footer>© Middags-Magi</footer> `;
+import { FrontPageView } from "./pages/FrontPageView.js";
+import { ExploreRecipesView } from "./pages/ExploreRecipesView.js";
+import { NewRecipeView } from "./pages/NewRecipeView.js";
+import { MyRecipesView } from "./pages/MyRecipesView.js";
+import { ShowRecipeView } from "./pages/ShowRecipeView.js";
+import { SuggestRecipeView } from "./pages/SuggestRecipeView.js";
+import { NavBarView } from "./ui/NavBarView.js";
+import { NavMenuView } from "./ui/NavMenuView.js";
+import { FooterView } from "./ui/FooterView.js";
 
 const page = () => {
     const pageHtml = pageManager();
-    return /* html */ ` ${navBar} ${navMenu} ${pageHtml} ${footer} `;
+    return /* html */ ` ${NavBarView()} ${NavMenuView()} ${pageHtml} ${FooterView()} `;
 };
 
 export function updateView() {
@@ -43,12 +32,12 @@ function pageManager() {
             return ExploreRecipesView();
         case "NewRecipe":
             return NewRecipeView();
-            case "MyRecipes":
-            return myRecipesView();
-        case "FullRecipe":
-            return FullRecipeView();
-        case "EditRecipe":
-            return editRecipeView();
+        case "MyRecipes":
+            return MyRecipesView();
+        case "ShowRecipe":
+            return ShowRecipeView();
+        case "SuggestRecipe":
+            return SuggestRecipeView();
         default:
             return "";
     }